refactor: place 관련 클래스를 workspace 패키지 하위로 이동
