+++ src/main/java/courseitda/workspace/infrastructure/JpaPlaceRepository.java
@@ -1,10 +1,8 @@
-package courseitda.place.infrastructure;
+package courseitda.workspace.infrastructure;
 
+import courseitda.workspace.domain.Place;
 import java.util.Optional;
-
 import org.springframework.data.jpa.repository.JpaRepository;
-
-import courseitda.place.domain.Place;
 
 public interface JpaPlaceRepository extends JpaRepository<Place, Long> {
