+++ src/test/java/courseitda/place/domain/PlaceBuilder.java
@@ -1,5 +1,7 @@
 package courseitda.place.domain;
 
+import courseitda.workspace.domain.Place;
+
 public class PlaceBuilder {
 
     private String name = PlaceFixture.anyName();
